refactor nostr.py
